improve the map/website links, add calendar events
The website and map links on the hotels, pre-party place and wedding
venue are now consistent, and they have a nice font icon in the front

Also I added a calendar event on the pre-party and venue. The calendar
event has both the pre-party and the wedding events included. In the
future I will add also the pick-up point.